third_party/blink/web_tests/fast/forms/checkbox/checkbox-nested-click-event-on-label-expected.txt

checkbox1 checkbox2

Tests nested click event on label element with checkbox

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.getElementById('ch2').checked is true
PASS successfullyParsed is true

TEST COMPLETE
